File rename for consistency: events_item -> event_item
