+++ app/lib/features/events/widgets/event_item.dart
